# packages/shared/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SprachPilot API",
    "version": "0.1.0",
    "description": "Shared Zod contracts drive validation and this OpenAPI document (SP-006)."
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Local API"
    }
  ],
  "components": {
    "schemas": {
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "maxLength": 500,
            "format": "uri"
          },
          "title": {
            "type": "string",
            "maxLength": 200
          },
          "status": {
            "type": "integer"
          },
          "detail": {
            "type": "string",
            "maxLength": 2000
          },
          "instance": {
            "type": "string",
            "maxLength": 500
          },
          "requestId": {
            "type": "string",
            "maxLength": 100
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldError"
            },
            "maxItems": 100
          }
        },
        "required": ["type", "title", "status"]
      },
      "FieldError": {
        "type": "object",
        "properties": {
          "field": {
            "type": "string",
            "maxLength": 200
          },
          "code": {
            "type": "string",
            "maxLength": 100
          },
          "message": {
            "type": "string",
            "maxLength": 500
          }
        },
        "required": ["field", "code", "message"]
      },
...
    },
...
  },
  "paths": {
...
    "/v1/auth/oauth/google": {
      "get": {
        "tags": ["Auth"],
        "summary": "Start Google OAuth",
        "description": "Redirects to Google with PKCE, state, and nonce. Behind FEATURE_GOOGLE_OAUTH and configured credentials.",
        "responses": {
          "302": {
            "description": "Redirect to Google authorization endpoint"
          },
          "403": {
            "description": "Google sign-in is not enabled.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "502": {
            "description": "Google sign-in is temporarily unavailable. Try again later.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Auth"],
        "summary": "Unlink Google",
        "description": "Removes the Google link when the account still has a password.",
        "responses": {
          "204": {
            "description": "Google unlinked"
          },
          "400": {
            "description": "Set a password before unlinking Google.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Missing or expired access token",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}